add halfshake to assoc id with cx/simplify mot

+++ include/client/opts.h
@@ -9,7 +9,7 @@ struct c_opts {
 	char ip_addr[OPT_STR_VALUE_LEN + 1];
 	char asset_path[OPT_STR_VALUE_LEN + 1];
 	uint8_t ui;
-	long id;
+	uint16_t id;
 };
 
 void process_c_opts(int argc, char * const *argv, struct c_opts *opts);
